Fix a few more issues left and add isolation_level to set.config()
